refactor: start of splitting up dto into sub packages
decided to go with grouping dtos into refs, summary, detail and response
this will be one of many commits before the code base is stable again

REFS #21
