+++ gallagher/dto/detail/__init__.py
@@ -1,0 +1,9 @@
+"""
+
+
+"""
+
+
+from .card_type import (
+    CardTypeDetail,
+)
