+++ gallagher/dto/ref/access_group.py
@@ -1,0 +1,13 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class AccessGroupRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ Access Groups is what a user is assigned to to provide access to doors
+    """
+    name: str
