+++ gallagher/dto/ref/alarm.py
@@ -1,0 +1,22 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin,
+)
+
+
+class AlarmRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ AlarmRef represents a single alarm
+    """
+    state: str
+
+
+class AlarmZoneRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ AccessZone represents
+    """
+    name: str
