+++ gallagher/dto/ref/cardholder.py
@@ -1,0 +1,13 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin,
+)
+
+
+class CardholderRef(
+    AppBaseModel,
+    HrefMixin,
+):
+    """ Reference to a Cardholder
+    """
+    name: str
