+++ gallagher/dto/response/alarm.py
@@ -1,0 +1,17 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin,
+)
+
+from ..summary import (
+    AlarmZoneSummary,
+)
+
+
+class AlarmResponse(
+    AppBaseModel,
+):
+    """ AlarmResponse represents a single alarm
+    """
+    alarms: list[AlarmZoneSummary]
+    updates: HrefMixin
