+++ gallagher/dto/response/division.py
@@ -1,0 +1,21 @@
+from typing import Optional
+
+from ..utils import (
+    AppBaseModel,
+    HrefMixin,
+)
+
+from ..detail import (
+    DivisionDetail
+)
+
+
+class DivisionDetailResponse(
+    AppBaseModel
+):
+    """ Division
+
+    """
+
+    results: list[DivisionDetail]
+    next: Optional[HrefMixin] = None
