+++ gallagher/dto/summary/card_type.py
@@ -1,0 +1,13 @@
+from typing import Optional
+
+from ..utils import (
+    AppBaseModel,
+    IdentityMixin,
+    HrefMixin
+)
+
+
+class CardExpiryTypeSummary(
+    AppBaseModel
+):
+    expiry_type: Optional[str] = None
